🧹 openai: split resource code into separate files and address PR review
Split monolithic openai.go into per-domain files for maintainability:
- openai.go: root resource, helpers (openaiConn, unixToTime, isAccessDenied)
- openai_model.go: models listing, isFineTuned, baseModel
- openai_file.go: files listing, initOpenaiFile
- openai_finetuning.go: fine-tuning jobs, Internal struct, trainingFile/validationFile
- openai_vectorstore.go: vector stores listing
- openai_project.go: projects, apiKeys, serviceAccounts
- openai_org.go: users, invites, auditLogs

PR review fixes:
- Use ListAutoPaging for models pagination
- Add 10s HTTP timeout to fetchAccountInfo
- Document /v1/me as undocumented endpoint
- Fix NEpochs union type handling (check OfAuto before OfInt)
- Limit audit logs to last 30 days via EffectiveAt.Gte filter
- Replace interface{} with any per linter
